Initial work to render using modern GL (shaders and VBO's)
Add a dependency on the cglm library, to do client-side linear algebra
instead of manipulating the OpenGL 1.1 matrix stack. Add code to
update the cglm matrices in the same places as the GL 1.1. matrix
stack is updated, and upload the modified MVP matrix as a uniform to
the GPU.

Update README with plans how to proceed.

Render top of dir/file boxes with VBO's and shaders in MAPV
mode. Remove use of OpenGL display lists in MAPV drawing as that
conflicts with VBO and shader rendering. Picking on these top surface
doesn't work, as that uses the deprecated glRenderMode(GL_SELECT)
which doesn't play together with VBO's and shaders. Follow-on work to
replace GL_SELECT with, say, an approach like
http://www.opengl-tutorial.org/miscellaneous/clicking-on-objects/picking-with-an-opengl-hack/