Implement Phong lightning model for splash and about screens
One reason why it looks off is a bug in the original code, namely that
normals for sharp corners in the sides are incorrect. These need
duplicate vertices with separate normals. Some commented out normal
debugging code in the shaders for working on this issue.

Now the letters F and S are rendered with modern GL, V with legacy.